MFC 1.41:
Try to close a potential, but serious race in our VM subsystem
that may result in panic with damaged page bucket hash table or
file system corruption.

The revision 1.41 tighten the check of vm_contigalloc1() and
vm_contigalloc2().

Please note that this MFC does not merged 1.39 and 1.40's changes.

Approved by:	alc